services/user-management-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.12.1
aiosqlite==0.19.0
greenlet==3.2.2
PyJWT==2.8.0
httpx==0.25.2
requests==2.31.0
email-validator==2.1.0
